+++ src/strspan.rs
@@ -1,16 +1,15 @@
 use core::fmt;
-use core::ops::Range;
+use core::ops::{Deref, Range};
 
 
-/// An immutable string slice.
+/// A string slice.
 ///
-/// Unlike `&str` contains a reference to the original string
-/// and a span region.
+/// Like `&str`, but also contains the position in the input XML,
+/// from which it was parsed.
 #[must_use]
 #[derive(Clone, Copy, PartialEq)]
 pub struct StrSpan<'a> {
     text: &'a str,
-    span: &'a str,
     start: usize,
 }
 
@@ -20,7 +19,6 @@ impl<'a> From<&'a str> for StrSpan<'a> {
         StrSpan {
             text,
             start: 0,
-            span: text,
         }
     }
 }
@@ -30,7 +28,7 @@ impl<'a> StrSpan<'a> {
     #[inline]
     pub(crate) fn from_substr(text: &str, start: usize, end: usize) -> StrSpan {
         debug_assert!(start <= end);
-        StrSpan { text, span: &text[start..end], start }
+        StrSpan { text: &text[start..end], start }
     }
 
     /// Returns a start position of the span.
@@ -42,7 +40,7 @@ impl<'a> StrSpan<'a> {
     /// Returns a end position of the span.
     #[inline]
     pub fn end(&self) -> usize {
-        self.start + self.span.len()
+        self.start + self.text.len()
     }
 
     /// Returns a end position of the span.
@@ -51,36 +49,15 @@ impl<'a> StrSpan<'a> {
         self.start..self.end()
     }
 
-    /// Returns a length of the span.
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.span.is_empty()
-    }
-
     /// Returns a span slice.
     #[inline]
     pub fn as_str(&self) -> &'a str {
-        &self.span
-    }
-
-    /// Returns a span slice as bytes.
-    #[inline]
-    pub(crate) fn as_bytes(&self) -> &'a [u8] {
-        self.span.as_bytes()
-    }
-
-    /// Returns an underling string.
-    #[inline]
-    pub fn full_str(&self) -> &'a str {
-        self.text
+        &self.text
     }
 
     /// Returns an underling string region as `StrSpan`.
     #[inline]
     pub(crate) fn slice_region(&self, start: usize, end: usize) -> StrSpan<'a> {
-        let start = self.start + start;
-        let end = self.start + end;
-
         StrSpan::from_substr(self.text, start, end)
     }
 }
@@ -94,5 +71,13 @@ impl<'a> fmt::Debug for StrSpan<'a> {
 impl<'a> fmt::Display for StrSpan<'a> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self.as_str())
+    }
+}
+
+impl<'a> Deref for StrSpan<'a> {
+    type Target = str;
+
+    fn deref(&self) -> &Self::Target {
+        self.text
     }
 }